checks fix

+++ Sources/Map.cpp
@@ -88,7 +88,7 @@ void Map::createMap() {
   startX += tileSize;
   startY += tileSize;
   endX =
-      static_cast<int>(vector[(int)vector.size() - 1]->rect.getPosition().x) -
+      vector[(int)vector.size() - 1]->rect.getPosition().x -
       tileSize;
   endY = static_cast<int>(vector[vector.size() - 1]->rect.getPosition().y) -
          tileSize;
@@ -142,7 +142,8 @@ void Map::checkPlayerCollision(Object *target) {
               vector[counter]->rect.getGlobalBounds().left +
                   vector[counter]->rect.getGlobalBounds().width,
               target->rect.getGlobalBounds().top);
-        else if (target->direction == -3 || target->direction == -4) {
+        else if (target->direction == -3 || target->direction == -4) 
+        {
           target->rect.setPosition(target->rect.getGlobalBounds().left,
                                    vector[counter]->rect.getGlobalBounds().top -
                                        target->rect.getGlobalBounds().height);
@@ -178,7 +179,8 @@ void Map::checkPlayerCollision(Object *target) {
               vector[counter]->rect.getGlobalBounds().left +
                   vector[counter]->rect.getGlobalBounds().width,
               target->rect.getGlobalBounds().top);
-        else if (target->direction == -1 || target->direction == -2) {
+        else if (target->direction == -1 || target->direction == -2) 
+        {
           target->rect.setPosition(
               target->rect.getGlobalBounds().left,
               vector[counter]->rect.getGlobalBounds().top +
@@ -263,7 +265,8 @@ void Map::checkBulletCollision(Object *target) {
       if (vector[counter]->type == BOX) {
         vector[counter]->hp--;
         target->isExist = false;
-        if (vector[counter]->hp <= 0) {
+        if (vector[counter]->hp <= 0) 
+        {
           vector[counter]->isExist = false;
         }
       } else if (vector[counter]->type == WALL)
